src/demo-edge-data-emitter/requirements.txt
aiomqtt==1.2.1
azure-iot-device==2.12.0
